fastboot: Extract common part of fastboot send functions
fastboot_fail(), fastboot_data(), fastboot_okay(), fastboot_info() and
fastboot_succeed() are very similar and can share the same code. Replece
these functions by thin wrappers to fastboot_send_fmt() function.

Change arguments for fastboot_fail() function to be printf like format
string.

based on:
commit 3d7592505385 ("fastboot: Extract common part of fastboot send
                      functions")

BRANCH=main
BUG=b:374092472
TEST=Build FW and run on Kano

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6574536

+++ src/fastboot/cmd.c
@@ -78,7 +78,7 @@ static void fastboot_cmd_upload(struct FastbootOps *fb, const char *arg)
 		return;
 	}
 
-	fastboot_data(fb, len);
+	fastboot_data(fb, (uint32_t)len);
 	fb->send_packet(fb, buf, len);
 	fastboot_succeed(fb);
